requirements.txt
pysocks
openai
tiktoken
graia-ariadne
pyyaml